app/requirements.txt
PySide6>=6.7
trimesh>=4.5
numpy>=1.26
shapely>=2.0
PyOpenGL>=3.1
